# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "authLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "authRegister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful"
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "summary": "User registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "authRefresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token refreshed successfully"
          },
          "401": {
            "description": "Invalid refresh token"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/buildings": {
      "post": {
        "operationId": "createBuilding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBuildingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Building created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new building",
        "tags": [
          "Buildings"
        ]
      },
      "get": {
        "operationId": "getBuildings",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "landlordId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of buildings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Building"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all buildings",
        "tags": [
          "Buildings"
        ]
      }
    },
    "/api/v1/buildings/{id}": {
      "get": {
        "operationId": "getBuilding",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Building ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Building found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "404": {
            "description": "Building not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get a building by ID",
        "tags": [
          "Buildings"
        ]
      },
      "patch": {
        "operationId": "updateBuilding",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Building ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBuildingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Building updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Building not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a building",
        "tags": [
          "Buildings"
        ]
      },
      "delete": {
        "operationId": "deleteBuilding",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Building ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Building deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Building not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a building",
        "tags": [
          "Buildings"
        ]
      }
    },
...
    "/api/v1/bills": {
      "post": {
        "operationId": "createBill",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBillDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bill created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new bill",
        "tags": [
          "Bills"
        ]
      },
      "get": {
        "operationId": "getBills",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "roomId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rentalId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PAID",
                "TENANT_CONFIRMED",
                "LANDLORD_CONFIRMED",
                "OVERDUE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of bills",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bill"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all bills",
        "tags": [
          "Bills"
        ]
      }
    },
    "/api/v1/bills/{id}": {
      "get": {
        "operationId": "getBill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Bill ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bill found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "404": {
            "description": "Bill not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get a bill by ID",
        "tags": [
          "Bills"
        ]
      },
      "patch": {
        "operationId": "updateBill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Bill ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBillDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bill updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Bill not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a bill",
        "tags": [
          "Bills"
        ]
      },
      "delete": {
        "operationId": "deleteBill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Bill ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bill deleted successfully"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Bill not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a bill",
        "tags": [
          "Bills"
        ]
      }
    },
    "/api/v1/bills/{id}/confirm": {
      "post": {
        "operationId": "confirmBillPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Bill ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment confirmed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bill"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Bill not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Confirm payment for a bill",
        "tags": [
          "Bills"
        ]
      }
    },
...
  },
  "info": {
    "title": "TacoHouse API",
    "description": "A comprehensive rental room management system API for managing buildings, rooms, tenants, billing, and payments.\n\n## Response Format\nAll responses follow a standardized format:\n- **status**: HTTP status code\n- **message**: Human-readable message\n- **data**: Response data (object or array)\n- **pagination**: (Optional) Pagination metadata for list endpoints",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
      "LoginAuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterAuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "LANDLORD",
              "TENANT"
            ]
          },
          "firstName": {
            "type": "string",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "maxLength": 10
          },
          "avatar": {
            "type": "string",
            "maxLength": 500
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "occupation": {
            "type": "string",
            "maxLength": 100
          },
          "workplace": {
            "type": "string",
            "maxLength": 100
          },
          "idCardFrontPhoto": {
            "type": "string",
            "maxLength": 500
          },
          "idCardBackPhoto": {
            "type": "string",
            "maxLength": 500
          },
          "portraitPhoto": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "email",
          "password",
          "role",
          "firstName",
          "lastName",
          "phone",
          "dateOfBirth",
          "occupation",
          "workplace"
        ]
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "LANDLORD",
          "TENANT"
        ]
      },
...
      "BillStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PAID",
          "TENANT_CONFIRMED",
          "LANDLORD_CONFIRMED",
          "OVERDUE"
        ]
      },
...
      "Bill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "billingPeriod": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "monthlyRent": {
            "type": "string",
            "format": "Decimal.js"
          },
          "electricityUsage": {
            "type": "string",
            "format": "Decimal.js"
          },
          "electricityAmount": {
            "type": "string",
            "format": "Decimal.js"
          },
          "waterUsage": {
            "type": "string",
            "format": "Decimal.js"
          },
          "waterAmount": {
            "type": "string",
            "format": "Decimal.js"
          },
          "gasUsage": {
            "type": "string",
            "format": "Decimal.js"
          },
          "gasAmount": {
            "type": "string",
            "format": "Decimal.js"
          },
          "managementFee": {
            "type": "string",
            "format": "Decimal.js"
          },
          "cleaningFee": {
            "type": "string",
            "format": "Decimal.js"
          },
          "lightingFee": {
            "type": "string",
            "format": "Decimal.js"
          },
          "previousDebt": {
            "type": "string",
            "format": "Decimal.js"
          },
          "totalAmount": {
            "type": "string",
            "format": "Decimal.js"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BillStatus"
              }
            ]
          },
          "payment": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Payment"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "roomId",
          "billingPeriod",
          "dueDate",
          "monthlyRent",
          "electricityUsage",
          "electricityAmount",
          "waterUsage",
          "waterAmount",
          "gasUsage",
          "gasAmount",
          "managementFee",
          "cleaningFee",
          "lightingFee",
          "previousDebt",
          "totalAmount",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ChatGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "buildingId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatGroupMember"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "buildingId",
          "name",
          "description",
          "createdAt",
          "updatedAt"
        ]
      },
      "Building": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "landlordId": {
            "type": "string"
          },
          "landlordRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "billingDate": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "electricityRate": {
            "type": "string",
            "format": "Decimal.js"
          },
          "waterRate": {
            "type": "string",
            "format": "Decimal.js"
          },
          "gasRate": {
            "type": "string",
            "format": "Decimal.js"
          },
          "managementFee": {
            "type": "string",
            "format": "Decimal.js"
          },
          "cleaningFeePerPerson": {
            "type": "string",
            "format": "Decimal.js"
          },
          "lightingFee": {
            "type": "string",
            "format": "Decimal.js"
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Room"
            }
          },
          "chatGroup": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatGroup"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "landlordId",
          "landlordRole",
          "name",
          "address",
          "description",
          "billingDate",
          "electricityRate",
          "waterRate",
          "gasRate",
          "managementFee",
          "cleaningFeePerPerson",
          "lightingFee",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateBuildingDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "billingDate": {
            "type": "number"
          },
          "landlordId": {
            "type": "string"
          },
          "electricityRate": {
            "type": "number"
          },
          "waterRate": {
            "type": "number"
          },
          "gasRate": {
            "type": "number"
          },
          "managementFee": {
            "type": "number"
          },
          "cleaningFeePerPerson": {
            "type": "number"
          },
          "lightingFee": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "address",
          "description",
          "landlordId"
        ]
      },
      "UpdateBuildingDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateBillDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string"
          },
          "billingPeriod": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "monthlyRent": {
            "type": "number",
            "minimum": 0
          },
          "electricityUsage": {
            "type": "number",
            "minimum": 0
          },
          "electricityAmount": {
            "type": "number",
            "minimum": 0
          },
          "waterUsage": {
            "type": "number",
            "minimum": 0
          },
          "waterAmount": {
            "type": "number",
            "minimum": 0
          },
          "gasUsage": {
            "type": "number",
            "minimum": 0
          },
          "gasAmount": {
            "type": "number",
            "minimum": 0
          },
          "managementFee": {
            "type": "number",
            "minimum": 0
          },
          "cleaningFee": {
            "type": "number",
            "minimum": 0
          },
          "lightingFee": {
            "type": "number",
            "minimum": 0
          },
          "previousDebt": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "roomId",
          "billingPeriod",
          "dueDate",
          "monthlyRent"
        ]
      },
      "UpdateBillDto": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "TENANT_CONFIRMED",
              "LANDLORD_CONFIRMED",
              "OVERDUE"
            ]
          }
        }
      },
      "ConfirmPaymentDto": {
        "type": "object",
        "properties": {
          "tenantConfirmed": {
            "type": "boolean"
          },
          "landlordConfirmed": {
            "type": "boolean"
          },
          "proofImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}